+++ R/sheets_sheet_delete.R
@@ -18,7 +18,7 @@
 #' @examples
 #' if (sheets_has_token()) {
 #'   ss <- sheets_create("delete-sheets-from-me")
-#'   sheets_sheet_add(ss, c("alpha", "beta", "gamma", "delta"))
+#'   sheet_add(ss, c("alpha", "beta", "gamma", "delta"))
 #'
 #'   # get an overview of the sheets
 #'   sheets_sheet_properties(ss)